Fold plain-line.js back into line.js
There's no need to have this in its own file, in fact it makes way more
sense to just have it in line.js.
