+++ js-lib/controller.js
@@ -2,7 +2,7 @@ import * as c from "./constants";
 import Vector from "./vector";
 import View from "./view";
 import State from "./state";
-import { DrawFunction, DrawBox, DrawLine, DrawPlainLine, DrawFreeform, DrawErase, DrawMove, DrawText, DrawSelect } from "./draw/index";
+import { DrawFunction, DrawBox, DrawLine, DrawFreeform, DrawErase, DrawMove, DrawText, DrawSelect } from "./draw/index";
 import DrawFunction from "./draw/function";
 
 /**
@@ -174,13 +174,13 @@ export default class Controller {
       this.drawFunction = new DrawBox(this.state);
     }
     if (id == "line-button") {
-      this.drawFunction = new DrawLine(this.state, false);
+      this.drawFunction = new DrawLine(this.state, 'connector');
     }
     if (id == "arrow-button") {
-      this.drawFunction = new DrawLine(this.state, true);
+      this.drawFunction = new DrawLine(this.state, 'arrow');
     }
     if (id == "plain-line-button") {
-      this.drawFunction = new DrawPlainLine(this.state);
+      this.drawFunction = new DrawLine(this.state, 'plain');
     }
     if (id == "freeform-button") {
       this.drawFunction = new DrawFreeform(this.state, "X");
