+++ js-lib/draw/index.js
@@ -2,7 +2,6 @@ export { default as DrawFunction } from "./function";
 export { default as DrawBox } from "./box";
 export { default as DrawErase } from "./erase";
 export { default as DrawLine } from "./line";
-export { default as DrawPlainLine } from "./plain-line";
 export { default as DrawSelect } from "./select";
 export { default as DrawText } from "./text";
 export { default as DrawMove } from "./move";
